# openapi.yaml
openapi: "3.1.0"
info:
  title: Lumenqraph API
  version: "0.1.0"
  description: |
    Open-source, self-hostable event indexer for Soroban smart contracts on Stellar.
    Exposes decoded contract events, materialized token transfers, on-chain
    contract interfaces, read-only view-function calls, and webhook management
    over a plain REST API.
  license:
    name: MIT
    url: https://github.com/Lumen-Scribe/Lumenqraph/blob/main/LICENSE
  contact:
    url: https://github.com/Lumen-Scribe/Lumenqraph
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyHeader: []
  - ApiKeyBearer: []
  - {}

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
    ApiKeyBearer:
      type: http
      scheme: bearer

  # ---------------------------------------------------------------------------
  # Shared schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    WebhookKind:
      type: string
      enum: [event, upgrade]
      description: |
        `event` — fires when a matching contract event is indexed.
        `upgrade` — fires when the contract's on-chain interface changes.
    CreateWebhookRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        kind:
          $ref: "#/components/schemas/WebhookKind"
        contract_id:
          type: string
          nullable: true
        event_name:
          type: string
          nullable: true
    WebhookSubscription:
      type: object
      required: [id, url, kind, active, created_at]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        kind:
          $ref: "#/components/schemas/WebhookKind"
        contract_id:
          type: string
          nullable: true
        event_name:
          type: string
          nullable: true
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    WebhookCreated:
      allOf:
        - $ref: "#/components/schemas/WebhookSubscription"
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              description: |
                HMAC-SHA256 signing secret. Returned only at creation — store it
                securely; it cannot be retrieved later. Use `verifyWebhook()` in
                the SDK to verify deliveries.

    # ---- health ----
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ---- pagination (#445) ----
...
  responses:
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /webhooks:
    post:
      operationId: createWebhook
      summary: Register a webhook subscription.
      tags: [webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "200":
          description: Subscription created. The `secret` field is only returned here.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookCreated"
    get:
      operationId: listWebhooks
      summary: List webhook subscriptions (secrets omitted).
      tags: [webhooks]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookSubscription"
  /webhooks/{id}:
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook subscription.
      tags: [webhooks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...